feat: beneficiary details screen (#2905)

+++ feature/beneficiary/src/commonMain/kotlin/org/mifos/mobile/feature/beneficiary/beneficiaryApplication/BeneficiaryApplicationScreen.kt
@@ -35,7 +35,7 @@ internal fun BeneficiaryApplicationScreen(
     navigateBack: () -> Unit,
     navigateToQR: () -> Unit,
     navigateToConfirmationScreen: (
-        beneficiaryId: Int,
+        beneficiaryId: Long,
         beneficiaryState: String,
         name: String,
         officeName: String,
